% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_data.R
\name{cluster_data}
\alias{cluster_data}
\title{Cluster Data}
\usage{
cluster_data(seurat_object, markers)
}
\arguments{
\item{seurat_object}{the object to retrieve the data from.}

\item{markers}{markers .}
}
\description{
This function computes the mean expression and number of positive cells for markers in the Seurat object.
}
\examples{
cluster_data()
}
\keyword{cluster}
