% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\docType{data}
\name{coordinates}
\alias{coordinates}
\title{Digiworm coordinates
The data with tSNE coordinates from digiworm. \n
Contains 4 elements:\n
1. all clusters
2. all neurons
3. ciliated neurons
4. non-ciliated neurons}
\format{An object of class \code{"list"}}
\source{
Omri Wurtzel
}
\usage{
data(coordinates)
}
\description{
Digiworm coordinates
The data with tSNE coordinates from digiworm. \n
Contains 4 elements:\n
1. all clusters
2. all neurons
3. ciliated neurons
4. non-ciliated neurons
}
\examples{
data(coordinates)
}
\references{
digiworm
}
\keyword{datasets}
