% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cells.R
\name{plot_cells}
\alias{plot_cells}
\title{Plot Cells}
\usage{
plot_cells(seurat_object, cells, alpha_neg = 0.1, reduction = "umap")
}
\arguments{
\item{seurat_object}{the object to retrieve the data from.}

\item{cells}{cells names (could be retrieved using \code{Seurat::Cells()})}

\item{alpha_neg}{transparency value for negative points.}

\item{reduction}{dimensionality reduction, could be \code{c('pca','tsne','umap')}}
}
\description{
This function plots the specific cells.
}
\examples{
plot_positives()
}
\keyword{cells}
\keyword{positive}
