% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cells_digiworm.R
\name{plot_cells_digiworm}
\alias{plot_cells_digiworm}
\title{Plot Cells in Digiworm}
\usage{
plot_cells_digiworm(cells, alpha_neg = 0.1)
}
\arguments{
\item{cells}{cells names (could be retrieved using \code{Seurat::Cells()})}

\item{alpha_neg}{transparency value for negative points.}
}
\description{
Plots specified cells across digiworm subsets (so far neuronal only)
}
\examples{
plot_cells_digiworm()
}
\keyword{cells}
\keyword{digiworm}
