% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Plot Cluster}
\usage{
plot_cluster(seurat_object, clusters, alpha_neg = 0.1,
  reduction = "umap")
}
\arguments{
\item{seurat_object}{the object to retrieve the data from.}

\item{clusters}{cluster or cluster to plot cells from.}

\item{alpha_neg}{transparency value for negative points.}

\item{reduction}{dimensionality reduction, could be \code{c('pca','tsne','umap')}}
}
\description{
This function plots a desired cluster contained in the seurat object
}
\examples{
plot_positives()
}
\keyword{cells}
\keyword{positive}
