% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne.R
\name{plot_digiworm}
\alias{plot_digiworm}
\title{Plot tSNE from digiworm}
\usage{
plot_digiworm(marker, alpha_neg)
}
\arguments{
\item{marker}{the marker to plot.}

\item{alpha_neg}{transparency value for negative points.}
}
\description{
This function plots a tSNE image for all subsets of cells in digiworm.
}
\examples{
plot_digiworm()
}
\keyword{cells}
\keyword{positive}
