% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positives.R
\name{plot_positives}
\alias{plot_positives}
\title{Plot Multiple Positives}
\usage{
plot_positives(seurat_object, markers, marker_names = NULL,
  alpha_neg = 0.1, reduction = "umap")
}
\arguments{
\item{seurat_object}{the object to retrieve the data from.}

\item{markers}{positive markers of the cells of interest.}

\item{marker_names}{optional names for markers}

\item{alpha_neg}{transparency value for negative points.}

\item{reduction}{dimensionality reduction, could be \code{c('pca','tsne','umap')}}
}
\description{
This function plots the cells that are positive at multiple markers at the same time.
}
\examples{
plot_positives()
}
\keyword{cells}
\keyword{positive}
