% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positives_digiworm.R
\name{plot_positives_digiworm}
\alias{plot_positives_digiworm}
\title{Plot Multiple Positives in Digiworm}
\usage{
plot_positives_digiworm(seurat_object, markers, marker_names = NULL,
  alpha_neg = 0.1)
}
\arguments{
\item{seurat_object}{the object to retrieve the cell data from ( cells that don't appear in this object are not plotted!)}

\item{markers}{positive markers of the cells of interest.}

\item{marker_names}{optional names for markers}

\item{alpha_neg}{transparency value for negative points.}
}
\description{
This function plots the cells that are positive at multiple markers at the same time using tSNE coordinatess from digiworm.
}
\examples{
plot_positives_digiworm()
}
\keyword{cells}
\keyword{positive}
