% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne.R
\name{plot_tsne}
\alias{plot_tsne}
\title{Plot tSNE from digiworm}
\usage{
plot_tsne(subset, marker, alpha_neg)
}
\arguments{
\item{subset}{the subset of cells from digiworm: 1 - all clusters; 2 - neural; 3 - ciliated; 4 - non-ciliated.}

\item{marker}{the marker to plot.}

\item{alpha_neg}{transparency value for negative points.}
}
\description{
This function plots a tSNE image from digiworm for a given marker
}
\examples{
plot_tsne()
}
\keyword{cells}
\keyword{positive}
