% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression.R
\name{plot_expression}
\alias{plot_expression}
\title{Plot Expression}
\usage{
plot_expression(seurat_object, marker, alpha_neg = 0.1,
  reduction = "umap")
}
\arguments{
\item{seurat_object}{the object to retrieve the data from.}

\item{marker}{a gene to plot.}

\item{alpha_neg}{transparency value for negative points.}

\item{reduction}{dimensionality reduction, could be \code{c('pca','tsne','umap')}}
}
\description{
This function plots the expression of the feature across cells. It's analogous to the \code{Seurat::FeaturePlot()} but it changes the transparency of negative cells so all positive ones are visible.
}
\examples{
cluster_data()
}
\keyword{expression}
